requirements.txt
anyio==4.0.0
APScheduler==3.10.4
certifi==2023.7.22
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
pdoc==14.1.0
pyaes==1.6.1
Pygments==2.16.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.0
python-telegram-bot==20.6
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
tzlocal==5.2
